Carriers/ Enzymes/ Targets/ Transporters Polypeptide parsers

Extract descriptions of identified polypeptide targets, enzymes, carriers, or transporters.

carriers_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

enzymes_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

targets_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

transporters_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

Arguments

save_table
boolean, save table in database if true.

save_csv
boolean, save csv version of parsed tibble if true

csv_path
location to save csv files into it, default is current location, save_csv must be true

override_csv
override existing csv, if any, in case it is true in the new parse operation

database_connection
DBI connection object that holds a connection to user defined database. If save_table is enabled without providing value for this function an error will be thrown.

Value

a tibble with 20 variables:

id

Universal Protein Resource (UniProt) identifier

source

Specifies whether the identified polypeptide ID is associated with any of the following UniProt knowledge bases: Swiss-Prot, which is manually annotated and reviewed, or TrEMBL, which is automatically annotated and not reviewed.

name

general_function

General summary of the physiological function of the polypeptide

specific_function

A more specific description of the polypeptide’s physiological function within the cell.

gene_name

The short name commonly associated with the associated gene. Eg. PTGS1.

locus

The specific chromosomal location or position of the gene’s sequence on a chromosome.

cellular_location

The cellular location of the polypeptide.

transmembrane_regions

Areas of the polypeptide sequence that span a biological membrane.

signal_regions

Location of any signal peptides within the polypeptide sequence.

theoretical_pi

Theoretical isoelectric point.

molecular_weight

The molecular weight of the polypeptide.

chromosome_location

The chromosomal location of the polypeptide gene

organism

The organism in which this polypeptide functions.

organism_ncbi_taxonomy_id

amino_acid_sequence

The amino acid sequence of the polypeptide

amino_acid_format

gene_sequence

The sequence of the associated gene.

gene_format

parent_key

carrier/ target/ enzyme/ transporter id

read_drugbank_xml_db

read_drugbank_xml_db function must be called first before any parser.

If read_drugbank_xml_db is called before for any reason, so no need to call it again before calling this function.

See also

Examples

if (FALSE) {
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

# will throw an error, as database_connection is NULL
run_all_parsers(save_table = TRUE)

# save in database in SQLite in memory database and return parsed tibble
sqlite_con <- DBI::dbConnect(RSQLite::SQLite())
run_all_parsers(save_table = TRUE, database_connection = sqlite_con)

# save parsed tibble as csv if it does not exist in current location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_csv = TRUE)

# save in database, save parsed tibble as csv,
# if it does not exist in current location and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_table = TRUE, save_csv = TRUE,
database_connection = sqlite_con)

# save parsed tibble as csv if it does not exist in given location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_csv = TRUE, csv_path = TRUE)

# save parsed tibble as csv if it does not exist in current location and
# return parsed tibble.
# if the csv exist override it and return it.
run_all_parsers(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}