Extracts the drug carriers polypeptides pfams element and return data as data frame.

parse_drug_carriers_polypeptides_pfams returns data frame of drug carriers polypeptides pfams elements.

parse_drug_carriers_polypeptides_pfams(save_table = FALSE,
  save_csv = FALSE, csv_path = ".", override_csv = FALSE)

Arguments

save_table
boolean, save table in database if true.

save_csv
boolean, save csv version of parsed dataframe if true

csv_path
location to save csv files into it, default is current location, save_csv must be true

override_csv
override existing csv, if any, in case it is true in the new parse operation

Value

drug carriers polypeptides pfams node attributes date frame

Details

This functions extracts the carriers polypeptides pfams element of drug node in drug bank xml database with the option to save it in a predefined database via open_db method. It takes one single optional argument to save the returned dataframe in the database. It must be called after get_xml_db_rows function like any other parser function. If get_xml_db_rows is called before for any reason, so no need to call it again before calling this function.

Examples

# return only the parsed dataframe
parse_drug_carriers_polypeptides_pfams()

#> # A tibble: 0 x 0

# save in database and return parsed dataframe
parse_drug_carriers_polypeptides_pfams(save_table = TRUE)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("NULL", "character", "tbl_df"), new("standardGeneric",     .Data = function (conn, name, value, ...)     standardGeneric("dbWriteTable"), generic = structure("dbWriteTable", package = "DBI"),     package = "DBI", group = list(), valueClass = character(0),     signature = c("conn", "name", "value"), default = NULL, skeleton = (function (conn,         name, value, ...)     stop("invalid call in method dispatch to 'dbWriteTable' (no default method)",         domain = NA))(conn, name, value, ...)), <environment>): unable to find an inherited method for function 'dbWriteTable' for signature '"NULL", "character", "tbl_df"'

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_carriers_polypeptides_pfams(save_csv = TRUE)

#> # A tibble: 0 x 0

# save in database, save parsed dataframe as csv if it does not exist in
#  current location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_carriers_polypeptides_pfams(ssave_table = TRUE, save_csv = TRUE)

#> Error in parse_drug_carriers_polypeptides_pfams(ssave_table = TRUE, save_csv = TRUE): unused argument (ssave_table = TRUE)

# save parsed dataframe as csv if it does not exist in given
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_carriers_polypeptides_pfams(save_csv = TRUE, csv_path = TRUE)

#> # A tibble: 0 x 0

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist override it and return it.
parse_drug_carriers_polypeptides_pfams(save_csv = TRUE, csv_path = TRUE, override = TRUE)

#> # A tibble: 0 x 0