Extracts the drug patents element and return data as tibble.

drug_patents returns tibble of drug patents elements.

drug_patents(save_table = FALSE, save_csv = FALSE, csv_path = ".",
  override_csv = FALSE)

Arguments

save_table
boolean, save table in database if true.

save_csv
boolean, save csv version of parsed dataframe if true

csv_path
location to save csv files into it, default is current location, save_csv must be true

override_csv
override existing csv, if any, in case it is true in the new parse operation

Value

drug patents node attributes tibble

Details

This functions extracts the patents element of drug node in drugbank xml database with the option to save it in a predefined database via open_db method. It takes one single optional argument to save the returned tibble in the database. It must be called after read_drugbank_xml_db function like any other parser function. If read_drugbank_xml_db is called before for any reason, so no need to call it again before calling this function.

See also

Examples

if (FALSE) {
# return only the parsed dataframe
drug_patents()

# save in database and return parsed dataframe
drug_patents(save_table = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
drug_patents(save_csv = TRUE)

# save in database, save parsed dataframe as csv if it does not exist
# 'in current location and return parsed dataframe.
# If the csv exist before read it and return its data.
drug_patents(save_table = TRUE, save_csv = TRUE)

# save parsed dataframe as csv if it does not exist in given
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
drug_patents(save_csv = TRUE, csv_path = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist override it and return it.
drug_patents(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}