extracts the all drug elements and return data as list of dataframes.

this functions extracts all element of drug nodes in drug bank xml database with the option to save it in a predefined database via open_db method. it takes one single optional argument to save the returned dataframe in the database. it must be called after get_xml_db_rows function like any other parser function. if get_xml_db_rows is called before for any reason, so no need to call it again before calling this function.

parse_drug_all(save_table = FALSE, save_csv = FALSE, csv_path = ".",
  override_csv = FALSE)

Arguments

save_table
boolean, save table in database if true.

save_csv
boolean, save csv version of parsed dataframe if true

csv_path
location to save csv files into it, default is current location, save_csv must be true

override_csv
override existing csv, if any, in case it is true in the new parse operation

Value

all drug elements dataframes

Examples

# \donttest{
# return only the parsed dataframe
parse_drug_all()

#> Warning: Missing column names filled in: 'X1' [1]

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )

#> parsed drugs main attributes, 1/75

#> parsed groups_drug, 2/75

#> parsed articles_drug, 3/75

#> parsed books_drug, 4/75

#> parsed links_drug, 5/75

#> parsed syn_drug, 6/75

#> parsed products_drug, 7/75

#> parsed mixtures_drug, 8/75

#> parsed packagers_drug, 9/75

#> parsed categories_drug, 10/75

#> parsed affected_organisms_drug, 11/75

#> parsed dosages_drug, 12/75

#> parsed ahfs_codes_drug, 13/75

#> parsed pdb_entries_drug, 14/75

#> parsed patents_drug, 15/75

#> parsed food_interactions_drug, 16/75

#> parsed interactions_drug, 17/75

#> parsed experimental_properties_drug, 18/75

#> parsed external_identifiers_drug, 19/75

#> parsed external_links_drug, 20/75

#> parsed snp_effects_drug, 21/75

#> parsed snp_adverse_reactions, 22/75

#> parsed atc_codes_drug, 23/75

#> parsed actions_carrier_drug, 24/75

#> parsed articles_carrier_drug, 25/75

#> parsed textbooks_carrier_drug, 26/75

#> parsed links_carrier_drug, 27/75

#> parsed polypeptides_carrier_drug, 28/75

#> parsed carr_poly_ext_identity, 29/75

#> parsed carr_polypeptides_syn, 30/75

#> parsed carr_polypeptides_pfams, 31/75

#> parsed carr_polypeptides_go, 32/75

#> parsed carriers_drug, 33/75

#> parsed classifications_drug, 34/75

#> parsed actions_enzyme_drug, 35/75

#> parsed articles_enzyme_drug, 36/75

#> parsed textbooks_enzyme_drug, 37/75

#> parsed links_enzyme_drug, 38/75

#> parsed polypeptides_enzyme_drug, 39/75

#> parsed external_identifiers_polypeptides_enzyme_drug, 40/75

#> parsed enzy_poly_syn, 41/75

#> parsed pfams_polypeptides_enzyme_drug, 42/75

#> parsed enzy_poly_go, 43/75

#> parsed enzyme_drug, 44/75

#> parsed manufacturers_drug, 45/75

#> parsed enzymes_pathway_drug, 46/75

#> parsed drug_pathway_drugs, 47/75

#> parsed pathways_drug, 48/75

#> parsed prices_drug, 49/75

#> parsed reactions_drug, 50/75

#> parsed enzymes_reactions_drug, 51/75

#> parsed sequences_drug, 52/75

#> parsed targ_poly_ext_identity, 53/75

#> parsed targ_poly_syn, 54/75

#> parsed pfams_polypeptide_target_drug, 55/75

#> parsed targ_poly_go attributes, 56/75

#> parsed actions_target_drug, 57/75

#> parsed articles_target_drug, 58/75

#> parsed textbooks_target_drug, 59/75

#> parsed links_target_drug, 60/75

#> parsed polypeptide_target_drug, 61/75

#> parsed targ_drug, 62/75

#> parsed actions_transporter_drug, 63/75

#> parsed articles_transporter_drug, 64/75

#> parsed textbooks_transporter_drug, 65/75

#> parsed links_transporter_drug, 66/75

#> parsed polypeptides_transporter_drug, 67/75

#> parsed trans_poly_ex_identity, 68/75

#> parsed syn_polypeptides_transporter_drug, 69/75

#> parsed pfams_polypeptides_transporter_drug, 70/75

#> parsed go_polypeptides_transporter_drug, 71/75

#> parsed transporters_drug, 72/75

#> parsed international_brands_drug, 73/75

#> parsed salts_drug, 74/75

#> parsed calculated_properties_drug, 75/75

#> $drugs
#> # A tibble: 0 x 1
#> # ... with 1 variable: X1 <chr>
#> 
#> $groups_drug
#> # A tibble: 0 x 0
#> 
#> $articles_drug
#> # A tibble: 0 x 0
#> 
#> $books_drug
#> # A tibble: 0 x 0
#> 
#> $links_drug
#> # A tibble: 0 x 0
#> 
#> $syn_drug
#> # A tibble: 0 x 0
#> 
#> $products_drug
#> # A tibble: 0 x 0
#> 
#> $mixtures_drug
#> # A tibble: 0 x 0
#> 
#> $packagers_drug
#> # A tibble: 0 x 0
#> 
#> $categories_drug
#> # A tibble: 0 x 0
#> 
#> $affected_organisms_drug
#> # A tibble: 0 x 0
#> 
#> $dosages_drug
#> # A tibble: 0 x 0
#> 
#> $ahfs_codes_drug
#> # A tibble: 0 x 0
#> 
#> $pdb_entries_drug
#> # A tibble: 0 x 0
#> 
#> $patents_drug
#> # A tibble: 0 x 0
#> 
#> $food_interactions_drug
#> # A tibble: 0 x 0
#> 
#> $interactions_drug
#> # A tibble: 0 x 0
#> 
#> $experimental_properties_drug
#> # A tibble: 0 x 0
#> 
#> $external_identifiers_drug
#> # A tibble: 0 x 0
#> 
#> $external_links_drug
#> # A tibble: 0 x 0
#> 
#> $snp_effects_drug
#> # A tibble: 0 x 0
#> 
#> $snp_adverse_reactions
#> # A tibble: 0 x 0
#> 
#> $atc_codes_drug
#> # A tibble: 0 x 0
#> 
#> $actions_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $articles_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $links_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptides_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $carr_poly_ext_identity
#> # A tibble: 0 x 0
#> 
#> $carr_polypeptides_syn
#> # A tibble: 0 x 0
#> 
#> $carr_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $carr_polypeptides_go
#> # A tibble: 0 x 0
#> 
#> $carriers_drug
#> # A tibble: 0 x 0
#> 
#> $classifications_drug
#> # A tibble: 0 x 0
#> 
#> $actions_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $articles_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $links_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptides_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $enzy_poly_ext_identity
#> # A tibble: 0 x 0
#> 
#> $enzy_poly_syn
#> # A tibble: 0 x 0
#> 
#> $pfams_polypeptides_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $enzy_poly_go
#> # A tibble: 0 x 0
#> 
#> $enzymes_drug
#> # A tibble: 0 x 0
#> 
#> $manufacturers_drug
#> # A tibble: 0 x 0
#> 
#> $enzymes_pathway_drug
#> # A tibble: 0 x 0
#> 
#> $drugs_pathway_drug
#> # A tibble: 0 x 0
#> 
#> $pathways_drug
#> # A tibble: 0 x 0
#> 
#> $prices_drug
#> # A tibble: 0 x 0
#> 
#> $reactions_drug
#> # A tibble: 0 x 0
#> 
#> $enzymes_reactions_drug
#> # A tibble: 0 x 0
#> 
#> $sequences_drug
#> # A tibble: 0 x 0
#> 
#> $targ_poly_ext_identity
#> # A tibble: 0 x 0
#> 
#> $targ_poly_syn
#> # A tibble: 0 x 0
#> 
#> $pfams_polypeptide_target_drug
#> # A tibble: 0 x 0
#> 
#> $targ_poly_go
#> # A tibble: 0 x 0
#> 
#> $actions_target_drug
#> # A tibble: 0 x 0
#> 
#> $articles_target_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_target_drug
#> # A tibble: 0 x 0
#> 
#> $links_target_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptide_target_drug
#> # A tibble: 0 x 0
#> 
#> $targ_drug
#> # A tibble: 0 x 0
#> 
#> $actions_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $articles_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $links_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptides_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $trans_poly_ex_identity
#> # A tibble: 0 x 0
#> 
#> $trans_poly_syn
#> # A tibble: 0 x 0
#> 
#> $trans_poly_pfams
#> # A tibble: 0 x 0
#> 
#> $trans_poly_go
#> # A tibble: 0 x 0
#> 
#> $transporters_drug
#> # A tibble: 0 x 0
#> 
#> $international_brands_drug
#> # A tibble: 0 x 0
#> 
#> $salts_drug
#> # A tibble: 0 x 0
#> 
#> $culculated_properties_drug
#> # A tibble: 0 x 0
#> 

# save in database and return parsed dataframe
parse_drug_all(save_table = trUe)

#> Warning: Missing column names filled in: 'X1' [1]

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )

#> Error in parse_drug(save_table, save_csv, csv_path, override_csv): object 'trUe' not found

# save parsed dataframe as csv if it does not exist in current location,
# and return parsed dataframe.
# if the csv exist before read it and return its data.
parse_drug_all(save_csv = trUe)

#> Warning: Missing column names filled in: 'X1' [1]

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )

#> parsed drugs main attributes, 1/75

#> parsed groups_drug, 2/75

#> parsed articles_drug, 3/75

#> parsed books_drug, 4/75

#> parsed links_drug, 5/75

#> parsed syn_drug, 6/75

#> parsed products_drug, 7/75

#> parsed mixtures_drug, 8/75

#> parsed packagers_drug, 9/75

#> parsed categories_drug, 10/75

#> parsed affected_organisms_drug, 11/75

#> parsed dosages_drug, 12/75

#> parsed ahfs_codes_drug, 13/75

#> parsed pdb_entries_drug, 14/75

#> parsed patents_drug, 15/75

#> parsed food_interactions_drug, 16/75

#> parsed interactions_drug, 17/75

#> parsed experimental_properties_drug, 18/75

#> parsed external_identifiers_drug, 19/75

#> parsed external_links_drug, 20/75

#> parsed snp_effects_drug, 21/75

#> parsed snp_adverse_reactions, 22/75

#> parsed atc_codes_drug, 23/75

#> parsed actions_carrier_drug, 24/75

#> parsed articles_carrier_drug, 25/75

#> parsed textbooks_carrier_drug, 26/75

#> parsed links_carrier_drug, 27/75

#> parsed polypeptides_carrier_drug, 28/75

#> parsed carr_poly_ext_identity, 29/75

#> parsed carr_polypeptides_syn, 30/75

#> parsed carr_polypeptides_pfams, 31/75

#> parsed carr_polypeptides_go, 32/75

#> parsed carriers_drug, 33/75

#> parsed classifications_drug, 34/75

#> parsed actions_enzyme_drug, 35/75

#> parsed articles_enzyme_drug, 36/75

#> parsed textbooks_enzyme_drug, 37/75

#> parsed links_enzyme_drug, 38/75

#> parsed polypeptides_enzyme_drug, 39/75

#> parsed external_identifiers_polypeptides_enzyme_drug, 40/75

#> parsed enzy_poly_syn, 41/75

#> parsed pfams_polypeptides_enzyme_drug, 42/75

#> parsed enzy_poly_go, 43/75

#> parsed enzyme_drug, 44/75

#> parsed manufacturers_drug, 45/75

#> parsed enzymes_pathway_drug, 46/75

#> parsed drug_pathway_drugs, 47/75

#> parsed pathways_drug, 48/75

#> parsed prices_drug, 49/75

#> parsed reactions_drug, 50/75

#> parsed enzymes_reactions_drug, 51/75

#> parsed sequences_drug, 52/75

#> parsed targ_poly_ext_identity, 53/75

#> parsed targ_poly_syn, 54/75

#> parsed pfams_polypeptide_target_drug, 55/75

#> parsed targ_poly_go attributes, 56/75

#> parsed actions_target_drug, 57/75

#> parsed articles_target_drug, 58/75

#> parsed textbooks_target_drug, 59/75

#> parsed links_target_drug, 60/75

#> parsed polypeptide_target_drug, 61/75

#> parsed targ_drug, 62/75

#> parsed actions_transporter_drug, 63/75

#> parsed articles_transporter_drug, 64/75

#> parsed textbooks_transporter_drug, 65/75

#> parsed links_transporter_drug, 66/75

#> parsed polypeptides_transporter_drug, 67/75

#> parsed trans_poly_ex_identity, 68/75

#> parsed syn_polypeptides_transporter_drug, 69/75

#> parsed pfams_polypeptides_transporter_drug, 70/75

#> parsed go_polypeptides_transporter_drug, 71/75

#> parsed transporters_drug, 72/75

#> parsed international_brands_drug, 73/75

#> parsed salts_drug, 74/75

#> parsed calculated_properties_drug, 75/75

#> $drugs
#> # A tibble: 0 x 1
#> # ... with 1 variable: X1 <chr>
#> 
#> $groups_drug
#> # A tibble: 0 x 0
#> 
#> $articles_drug
#> # A tibble: 0 x 0
#> 
#> $books_drug
#> # A tibble: 0 x 0
#> 
#> $links_drug
#> # A tibble: 0 x 0
#> 
#> $syn_drug
#> # A tibble: 0 x 0
#> 
#> $products_drug
#> # A tibble: 0 x 0
#> 
#> $mixtures_drug
#> # A tibble: 0 x 0
#> 
#> $packagers_drug
#> # A tibble: 0 x 0
#> 
#> $categories_drug
#> # A tibble: 0 x 0
#> 
#> $affected_organisms_drug
#> # A tibble: 0 x 0
#> 
#> $dosages_drug
#> # A tibble: 0 x 0
#> 
#> $ahfs_codes_drug
#> # A tibble: 0 x 0
#> 
#> $pdb_entries_drug
#> # A tibble: 0 x 0
#> 
#> $patents_drug
#> # A tibble: 0 x 0
#> 
#> $food_interactions_drug
#> # A tibble: 0 x 0
#> 
#> $interactions_drug
#> # A tibble: 0 x 0
#> 
#> $experimental_properties_drug
#> # A tibble: 0 x 0
#> 
#> $external_identifiers_drug
#> # A tibble: 0 x 0
#> 
#> $external_links_drug
#> # A tibble: 0 x 0
#> 
#> $snp_effects_drug
#> # A tibble: 0 x 0
#> 
#> $snp_adverse_reactions
#> # A tibble: 0 x 0
#> 
#> $atc_codes_drug
#> # A tibble: 0 x 0
#> 
#> $actions_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $articles_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $links_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptides_carrier_drug
#> # A tibble: 0 x 0
#> 
#> $carr_poly_ext_identity
#> # A tibble: 0 x 0
#> 
#> $carr_polypeptides_syn
#> # A tibble: 0 x 0
#> 
#> $carr_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $carr_polypeptides_go
#> # A tibble: 0 x 0
#> 
#> $carriers_drug
#> # A tibble: 0 x 0
#> 
#> $classifications_drug
#> # A tibble: 0 x 0
#> 
#> $actions_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $articles_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $links_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptides_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $enzy_poly_ext_identity
#> # A tibble: 0 x 0
#> 
#> $enzy_poly_syn
#> # A tibble: 0 x 0
#> 
#> $pfams_polypeptides_enzyme_drug
#> # A tibble: 0 x 0
#> 
#> $enzy_poly_go
#> # A tibble: 0 x 0
#> 
#> $enzymes_drug
#> # A tibble: 0 x 0
#> 
#> $manufacturers_drug
#> # A tibble: 0 x 0
#> 
#> $enzymes_pathway_drug
#> # A tibble: 0 x 0
#> 
#> $drugs_pathway_drug
#> # A tibble: 0 x 0
#> 
#> $pathways_drug
#> # A tibble: 0 x 0
#> 
#> $prices_drug
#> # A tibble: 0 x 0
#> 
#> $reactions_drug
#> # A tibble: 0 x 0
#> 
#> $enzymes_reactions_drug
#> # A tibble: 0 x 0
#> 
#> $sequences_drug
#> # A tibble: 0 x 0
#> 
#> $targ_poly_ext_identity
#> # A tibble: 0 x 0
#> 
#> $targ_poly_syn
#> # A tibble: 0 x 0
#> 
#> $pfams_polypeptide_target_drug
#> # A tibble: 0 x 0
#> 
#> $targ_poly_go
#> # A tibble: 0 x 0
#> 
#> $actions_target_drug
#> # A tibble: 0 x 0
#> 
#> $articles_target_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_target_drug
#> # A tibble: 0 x 0
#> 
#> $links_target_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptide_target_drug
#> # A tibble: 0 x 0
#> 
#> $targ_drug
#> # A tibble: 0 x 0
#> 
#> $actions_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $articles_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $textbooks_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $links_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $polypeptides_transporter_drug
#> # A tibble: 0 x 0
#> 
#> $trans_poly_ex_identity
#> # A tibble: 0 x 0
#> 
#> $trans_poly_syn
#> # A tibble: 0 x 0
#> 
#> $trans_poly_pfams
#> # A tibble: 0 x 0
#> 
#> $trans_poly_go
#> # A tibble: 0 x 0
#> 
#> $transporters_drug
#> # A tibble: 0 x 0
#> 
#> $international_brands_drug
#> # A tibble: 0 x 0
#> 
#> $salts_drug
#> # A tibble: 0 x 0
#> 
#> $culculated_properties_drug
#> # A tibble: 0 x 0
#> 

# save in database, save parsed dataframe as csv,
# if it does not exist in current location and return parsed dataframe.
# if the csv exist before read it and return its data.
parse_drug_all(ssave_table = trUe, save_csv = trUe)

#> Error in parse_drug_all(ssave_table = trUe, save_csv = trUe): unused argument (ssave_table = trUe)

# save parsed dataframe as csv if it does not exist in given location,
# and return parsed dataframe.
# if the csv exist before read it and return its data.
parse_drug_all(save_csv = trUe, csv_path = trUe)

#> Error in ifelse(csv_path == ".", file.path(getwd(), paste0(data, ".csv")),     file.path(csv_path, paste0(data, ".csv"))): object 'trUe' not found

# save parsed dataframe as csv if it does not exist in current location and
# return parsed dataframe.
# if the csv exist override it and return it.
parse_drug_all(save_csv = trUe, csv_path = trUe, override = trUe)

#> Error in ifelse(csv_path == ".", file.path(getwd(), paste0(data, ".csv")),     file.path(csv_path, paste0(data, ".csv"))): object 'trUe' not found

# }