Extracts the all drug elements and return data as list of dataframes.

This functions extracts all element of drug nodes in drug bank xml database with the option to save it in a predefined database via open_db method. It takes one single optional argument to save the returned dataframe in the database. It must be called after get_xml_db_rows function like any other parser function. If get_xml_db_rows is called before for any reason, so no need to call it again before calling this function.

parse_drug_all(save_table = FALSE, save_csv = FALSE, csv_path = ".",
  override_csv = FALSE)

Arguments

save_table
boolean, save table in database if true.

save_csv
boolean, save csv version of parsed dataframe if true

csv_path
location to save csv files into it, default is current location, save_csv must be true

override_csv
override existing csv, if any, in case it is true in the new parse operation

Value

all drug elements dataframes

Examples

# \donttest{
# return only the parsed dataframe
parse_drug_all()

#> Warning: Missing column names filled in: 'X1' [1]

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )

#> [1] "Parsed Drugs main attributes, 1/75"
#> [1] "Parsed Groups_Drug, 2/75"
#> [1] "Parsed Articles_Drug, 3/75"
#> [1] "Parsed Books_Drug, 4/75"
#> [1] "Parsed Links_Drug, 5/75"
#> [1] "Parsed Synonyms_Drug, 6/75"
#> [1] "Parsed Products_Drug, 7/75"
#> [1] "Parsed Mixtures_Drug, 8/75"
#> [1] "Parsed Packagers_Drug, 9/75"
#> [1] "Parsed Categories_Drug, 10/75"
#> [1] "Parsed Affected_Organisms_Drug, 11/75"
#> [1] "Parsed Dosages_Drug, 12/75"
#> [1] "Parsed AHFS_Codes_Drug, 13/75"
#> [1] "Parsed PDB_Entries_Drug, 14/75"
#> [1] "Parsed Patents_Drug, 15/75"
#> [1] "Parsed Food_Interactions_Drug, 16/75"
#> [1] "Parsed Interactions_Drug, 17/75"
#> [1] "Parsed Experimental_Properties_Drug, 18/75"
#> [1] "Parsed External_Identifiers_Drug, 19/75"
#> [1] "Parsed External_Links_Drug, 20/75"
#> [1] "Parsed SNP_Effects_Drug, 21/75"
#> [1] "Parsed SNP_Adverse_Drug_Reactions_Drug, 22/75"
#> [1] "Parsed ATC_Codes_Drug, 23/75"
#> [1] "Parsed Actions_Carrier_Drug, 24/75"
#> [1] "Parsed Articles_Carrier_Drug, 25/75"
#> [1] "Parsed Textbooks_Carrier_Drug, 26/75"
#> [1] "Parsed Links_Carrier_Drug, 27/75"
#> [1] "Parsed Polypeptides_Carrier_Drug, 28/75"
#> [1] "Parsed External_Identifiers_Polypeptide_Carrier_Drug, 29/75"
#> [1] "Parsed Synonyms_Polypeptide_Carrier_Drug, 30/75"
#> [1] "Parsed PFAMS_Polypeptide_Carrier_Drug, 31/75"
#> [1] "Parsed GO_Classifiers_Polypeptide_Carrier_Drug, 32/75"
#> [1] "Parsed Carriers_Drug, 33/75"
#> [1] "Parsed Classifications_Drug, 34/75"
#> [1] "Parsed Actions_Enzyme_Drug, 35/75"
#> [1] "Parsed Articles_Enzyme_Drug, 36/75"
#> [1] "Parsed Textbooks_Enzyme_Drug, 37/75"
#> [1] "Parsed Links_Enzyme_Drug, 38/75"
#> [1] "Parsed Polypeptides_Enzyme_Drug, 39/75"
#> [1] "Parsed External_Identifiers_Polypeptides_Enzyme_Drug, 40/75"
#> [1] "Parsed Synonyms_Polypeptides_Enzyme_Drug, 41/75"
#> [1] "Parsed PFAMS_Polypeptides_Enzyme_Drug, 42/75"
#> [1] "Parsed GO_Classifiers_Polypeptides_Enzyme_Drug, 43/75"
#> [1] "Parsed Enzyme_Drug, 44/75"
#> [1] "Parsed Manufacturers_Drug, 45/75"
#> [1] "Parsed Enzymes_Pathway_Drug, 46/75"
#> [1] "Parsed drug_pathway_drugs, 47/75"
#> [1] "Parsed Pathways_Drug, 48/75"
#> [1] "Parsed Prices_Drug, 49/75"
#> [1] "Parsed Reactions_Drug, 50/75"
#> [1] "Parsed Enzymes_Reactions_Drug, 51/75"
#> [1] "Parsed Sequences_Drug, 52/75"
#> [1] "Parsed External_Identifiers_Polypeptide_Target_Drug, 53/75"
#> [1] "Parsed Synonyms_Polypeptide_Target_Drug, 54/75"
#> [1] "Parsed PFAMS_Polypeptide_Target_Drug, 55/75"
#> [1] "Parsed GO_Classifiers_Polypeptide_Target_Drug attributes, 56/75"
#> [1] "Parsed Actions_Target_Drug, 57/75"
#> [1] "Parsed Articles_Target_Drug, 58/75"
#> [1] "Parsed Textbooks_Target_Drug, 59/75"
#> [1] "Parsed Links_Target_Drug, 60/75"
#> [1] "Parsed Polypeptide_Target_Drug, 61/75"
#> [1] "Parsed Targets_Drug, 62/75"
#> [1] "Parsed Actions_Transporter_Drug, 63/75"
#> [1] "Parsed Articles_Transporter_Drug, 64/75"
#> [1] "Parsed Textbooks_Transporter_Drug, 65/75"
#> [1] "Parsed Links_Transporter_Drug, 66/75"
#> [1] "Parsed Polypeptides_Transporter_Drug, 67/75"
#> [1] "Parsed External_Identifiers_Transporter_Drug, 68/75"
#> [1] "Parsed Synonyms_Polypeptides_Transporter_Drug, 69/75"
#> [1] "Parsed PFAMS_Polypeptides_Transporter_Drug, 70/75"
#> [1] "Parsed GO_Classifiers_Polypeptides_Transporter_Drug, 71/75"
#> [1] "Parsed Transporters_Drug, 72/75"
#> [1] "Parsed International_Brands_Drug, 73/75"
#> [1] "Parsed Salts_Drug, 74/75"
#> [1] "Parsed Calculated_Properties_Drug, 75/75"

#> $Drugs
#> # A tibble: 0 x 1
#> # ... with 1 variable: X1 <chr>
#> 
#> $Groups_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Drug
#> # A tibble: 0 x 0
#> 
#> $Books_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Drug
#> # A tibble: 0 x 0
#> 
#> $Products_Drug
#> # A tibble: 0 x 0
#> 
#> $Mixtures_Drug
#> # A tibble: 0 x 0
#> 
#> $Packagers_Drug
#> # A tibble: 0 x 0
#> 
#> $Categories_Drug
#> # A tibble: 0 x 0
#> 
#> $Affected_Organisms_Drug
#> # A tibble: 0 x 0
#> 
#> $Dosages_Drug
#> # A tibble: 0 x 0
#> 
#> $AHFS_Codes_Drug
#> # A tibble: 0 x 0
#> 
#> $PDB_Entries_Drug
#> # A tibble: 0 x 0
#> 
#> $Patents_Drug
#> # A tibble: 0 x 0
#> 
#> $Food_Interactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Interactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Experimental_Properties_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Links_Drug
#> # A tibble: 0 x 0
#> 
#> $SNP_Effects_Drug
#> # A tibble: 0 x 0
#> 
#> $SNP_Adverse_Drug_Reactions_Drug
#> # A tibble: 0 x 0
#> 
#> $ATC_Codes_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptides_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Carriers_Drug
#> # A tibble: 0 x 0
#> 
#> $Classifications_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Polypeptide_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Enzymes_Drug
#> # A tibble: 0 x 0
#> 
#> $Manufacturers_Drug
#> # A tibble: 0 x 0
#> 
#> $Enzymes_Pathway_Drug
#> # A tibble: 0 x 0
#> 
#> $Drugs_Pathway_Drug
#> # A tibble: 0 x 0
#> 
#> $Pathways_Drug
#> # A tibble: 0 x 0
#> 
#> $Prices_Drug
#> # A tibble: 0 x 0
#> 
#> $Reactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Enzymes_Reactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Sequences_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Targets_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptides_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptide_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptid_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptide_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Transporters_Drug
#> # A tibble: 0 x 0
#> 
#> $International_Brands_Drug
#> # A tibble: 0 x 0
#> 
#> $Salts_Drug
#> # A tibble: 0 x 0
#> 
#> $Culculated_Properties_Drug
#> # A tibble: 0 x 0
#> 

# save in database and return parsed dataframe
parse_drug_all(save_table = TRUE)

#> Warning: Missing column names filled in: 'X1' [1]

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("NULL", "character", "spec_tbl_df"), new("standardGeneric",     .Data = function (conn, name, value, ...)     standardGeneric("dbWriteTable"), generic = structure("dbWriteTable", package = "DBI"),     package = "DBI", group = list(), valueClass = character(0),     signature = c("conn", "name", "value"), default = NULL, skeleton = (function (conn,         name, value, ...)     stop("invalid call in method dispatch to 'dbWriteTable' (no default method)",         domain = NA))(conn, name, value, ...)), <environment>): unable to find an inherited method for function 'dbWriteTable' for signature '"NULL", "character", "spec_tbl_df"'

# save parsed dataframe as csv if it does not exist in current location,
# and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_all(save_csv = TRUE)

#> Warning: Missing column names filled in: 'X1' [1]

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )

#> [1] "Parsed Drugs main attributes, 1/75"
#> [1] "Parsed Groups_Drug, 2/75"
#> [1] "Parsed Articles_Drug, 3/75"
#> [1] "Parsed Books_Drug, 4/75"
#> [1] "Parsed Links_Drug, 5/75"
#> [1] "Parsed Synonyms_Drug, 6/75"
#> [1] "Parsed Products_Drug, 7/75"
#> [1] "Parsed Mixtures_Drug, 8/75"
#> [1] "Parsed Packagers_Drug, 9/75"
#> [1] "Parsed Categories_Drug, 10/75"
#> [1] "Parsed Affected_Organisms_Drug, 11/75"
#> [1] "Parsed Dosages_Drug, 12/75"
#> [1] "Parsed AHFS_Codes_Drug, 13/75"
#> [1] "Parsed PDB_Entries_Drug, 14/75"
#> [1] "Parsed Patents_Drug, 15/75"
#> [1] "Parsed Food_Interactions_Drug, 16/75"
#> [1] "Parsed Interactions_Drug, 17/75"
#> [1] "Parsed Experimental_Properties_Drug, 18/75"
#> [1] "Parsed External_Identifiers_Drug, 19/75"
#> [1] "Parsed External_Links_Drug, 20/75"
#> [1] "Parsed SNP_Effects_Drug, 21/75"
#> [1] "Parsed SNP_Adverse_Drug_Reactions_Drug, 22/75"
#> [1] "Parsed ATC_Codes_Drug, 23/75"
#> [1] "Parsed Actions_Carrier_Drug, 24/75"
#> [1] "Parsed Articles_Carrier_Drug, 25/75"
#> [1] "Parsed Textbooks_Carrier_Drug, 26/75"
#> [1] "Parsed Links_Carrier_Drug, 27/75"
#> [1] "Parsed Polypeptides_Carrier_Drug, 28/75"
#> [1] "Parsed External_Identifiers_Polypeptide_Carrier_Drug, 29/75"
#> [1] "Parsed Synonyms_Polypeptide_Carrier_Drug, 30/75"
#> [1] "Parsed PFAMS_Polypeptide_Carrier_Drug, 31/75"
#> [1] "Parsed GO_Classifiers_Polypeptide_Carrier_Drug, 32/75"
#> [1] "Parsed Carriers_Drug, 33/75"
#> [1] "Parsed Classifications_Drug, 34/75"
#> [1] "Parsed Actions_Enzyme_Drug, 35/75"
#> [1] "Parsed Articles_Enzyme_Drug, 36/75"
#> [1] "Parsed Textbooks_Enzyme_Drug, 37/75"
#> [1] "Parsed Links_Enzyme_Drug, 38/75"
#> [1] "Parsed Polypeptides_Enzyme_Drug, 39/75"
#> [1] "Parsed External_Identifiers_Polypeptides_Enzyme_Drug, 40/75"
#> [1] "Parsed Synonyms_Polypeptides_Enzyme_Drug, 41/75"
#> [1] "Parsed PFAMS_Polypeptides_Enzyme_Drug, 42/75"
#> [1] "Parsed GO_Classifiers_Polypeptides_Enzyme_Drug, 43/75"
#> [1] "Parsed Enzyme_Drug, 44/75"
#> [1] "Parsed Manufacturers_Drug, 45/75"
#> [1] "Parsed Enzymes_Pathway_Drug, 46/75"
#> [1] "Parsed drug_pathway_drugs, 47/75"
#> [1] "Parsed Pathways_Drug, 48/75"
#> [1] "Parsed Prices_Drug, 49/75"
#> [1] "Parsed Reactions_Drug, 50/75"
#> [1] "Parsed Enzymes_Reactions_Drug, 51/75"
#> [1] "Parsed Sequences_Drug, 52/75"
#> [1] "Parsed External_Identifiers_Polypeptide_Target_Drug, 53/75"
#> [1] "Parsed Synonyms_Polypeptide_Target_Drug, 54/75"
#> [1] "Parsed PFAMS_Polypeptide_Target_Drug, 55/75"
#> [1] "Parsed GO_Classifiers_Polypeptide_Target_Drug attributes, 56/75"
#> [1] "Parsed Actions_Target_Drug, 57/75"
#> [1] "Parsed Articles_Target_Drug, 58/75"
#> [1] "Parsed Textbooks_Target_Drug, 59/75"
#> [1] "Parsed Links_Target_Drug, 60/75"
#> [1] "Parsed Polypeptide_Target_Drug, 61/75"
#> [1] "Parsed Targets_Drug, 62/75"
#> [1] "Parsed Actions_Transporter_Drug, 63/75"
#> [1] "Parsed Articles_Transporter_Drug, 64/75"
#> [1] "Parsed Textbooks_Transporter_Drug, 65/75"
#> [1] "Parsed Links_Transporter_Drug, 66/75"
#> [1] "Parsed Polypeptides_Transporter_Drug, 67/75"
#> [1] "Parsed External_Identifiers_Transporter_Drug, 68/75"
#> [1] "Parsed Synonyms_Polypeptides_Transporter_Drug, 69/75"
#> [1] "Parsed PFAMS_Polypeptides_Transporter_Drug, 70/75"
#> [1] "Parsed GO_Classifiers_Polypeptides_Transporter_Drug, 71/75"
#> [1] "Parsed Transporters_Drug, 72/75"
#> [1] "Parsed International_Brands_Drug, 73/75"
#> [1] "Parsed Salts_Drug, 74/75"
#> [1] "Parsed Calculated_Properties_Drug, 75/75"

#> $Drugs
#> # A tibble: 0 x 1
#> # ... with 1 variable: X1 <chr>
#> 
#> $Groups_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Drug
#> # A tibble: 0 x 0
#> 
#> $Books_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Drug
#> # A tibble: 0 x 0
#> 
#> $Products_Drug
#> # A tibble: 0 x 0
#> 
#> $Mixtures_Drug
#> # A tibble: 0 x 0
#> 
#> $Packagers_Drug
#> # A tibble: 0 x 0
#> 
#> $Categories_Drug
#> # A tibble: 0 x 0
#> 
#> $Affected_Organisms_Drug
#> # A tibble: 0 x 0
#> 
#> $Dosages_Drug
#> # A tibble: 0 x 0
#> 
#> $AHFS_Codes_Drug
#> # A tibble: 0 x 0
#> 
#> $PDB_Entries_Drug
#> # A tibble: 0 x 0
#> 
#> $Patents_Drug
#> # A tibble: 0 x 0
#> 
#> $Food_Interactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Interactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Experimental_Properties_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Links_Drug
#> # A tibble: 0 x 0
#> 
#> $SNP_Effects_Drug
#> # A tibble: 0 x 0
#> 
#> $SNP_Adverse_Drug_Reactions_Drug
#> # A tibble: 0 x 0
#> 
#> $ATC_Codes_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptides_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptide_Carrier_Drug
#> # A tibble: 0 x 0
#> 
#> $Carriers_Drug
#> # A tibble: 0 x 0
#> 
#> $Classifications_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Polypeptide_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptides_Enzyme_Drug
#> # A tibble: 0 x 0
#> 
#> $Enzymes_Drug
#> # A tibble: 0 x 0
#> 
#> $Manufacturers_Drug
#> # A tibble: 0 x 0
#> 
#> $Enzymes_Pathway_Drug
#> # A tibble: 0 x 0
#> 
#> $Drugs_Pathway_Drug
#> # A tibble: 0 x 0
#> 
#> $Pathways_Drug
#> # A tibble: 0 x 0
#> 
#> $Prices_Drug
#> # A tibble: 0 x 0
#> 
#> $Reactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Enzymes_Reactions_Drug
#> # A tibble: 0 x 0
#> 
#> $Sequences_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptide_Target_Drug
#> # A tibble: 0 x 0
#> 
#> $Targets_Drug
#> # A tibble: 0 x 0
#> 
#> $Actions_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Articles_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Textbooks_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Links_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Polypeptides_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $External_Identifiers_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Synonyms_Polypeptide_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $PFAMS_Polypeptid_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $GO_Classifiers_Polypeptide_Transporter_Drug
#> # A tibble: 0 x 0
#> 
#> $Transporters_Drug
#> # A tibble: 0 x 0
#> 
#> $International_Brands_Drug
#> # A tibble: 0 x 0
#> 
#> $Salts_Drug
#> # A tibble: 0 x 0
#> 
#> $Culculated_Properties_Drug
#> # A tibble: 0 x 0
#> 

# save in database, save parsed dataframe as csv,
# if it does not exist in current location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_all(ssave_table = TRUE, save_csv = TRUE)

#> Error in parse_drug_all(ssave_table = TRUE, save_csv = TRUE): unused argument (ssave_table = TRUE)

# save parsed dataframe as csv if it does not exist in given location,
# and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_all(save_csv = TRUE, csv_path = TRUE)

#> Warning: cannot open file 'TRUE/drugs.csv': No such file or directory

#> Error in open.connection(path, "wb"): cannot open the connection

# save parsed dataframe as csv if it does not exist in current location and return parsed dataframe.
# If the csv exist override it and return it.
parse_drug_all(save_csv = TRUE, csv_path = TRUE, override = TRUE)

#> Warning: cannot open file 'TRUE/drugs.csv': No such file or directory

#> Error in open.connection(path, "wb"): cannot open the connection

# }