Extracts the all drug elements and return data as list of dataframes.

parse_drug_all returns list of dataframes of drugs elements.

parse_drug_all(save_table = FALSE)

Arguments

save_table
boolean, save table in database if true.

Value

all drug elements dataframes

Details

This functions extracts all element of drug nodes in drug bank xml database with the option to save it in a predefined database via open_db method. It takes one single optional argument to save the returned dataframe in the database. It must be called after get_xml_db_rows function like any other parser function. If get_xml_db_rows is called before for any reason, so no need to call it again before calling this function.

Examples

parse_drug_all()

#> [1] "Parsed drugs main attributes, 1/72"
#> [1] "Parsed drug_groups, 2/72"
#> [1] "Parsed drug_articles, 3/72"
#> [1] "Parsed drug_books, 4/72"
#> [1] "Parsed drug_links, 5/72"
#> [1] "Parsed drug_synonyms, 6/72"
#> [1] "Parsed drug_products, 7/72"
#> [1] "Parsed drug_mixtures, 8/72"
#> [1] "Parsed drug_packagers, 9/72"
#> [1] "Parsed drug_categories, 10/72"
#> [1] "Parsed drug_affected_organisms, 11/72"
#> [1] "Parsed drug_dosages, 12/72"
#> [1] "Parsed drug_ahfs_codes, 13/72"
#> [1] "Parsed drug_pdb_entries, 14/72"
#> [1] "Parsed drug_patents, 15/72"
#> [1] "Parsed drug_food_interactions, 16/72"
#> [1] "Parsed drug_interactions, 17/72"
#> [1] "Parsed drug_experimental_properties, 18/72"
#> [1] "Parsed drug_external_identifiers, 19/72"
#> [1] "Parsed drug_external_links, 20/72"
#> [1] "Parsed drug_snp_effects, 21/72"
#> [1] "Parsed drug_snp_adverse_drug_reactions, 22/72"
#> [1] "Parsed drug_atc_codes, 23/72"
#> [1] "Parsed drug_carriers_actions, 24/72"
#> [1] "Parsed drug_carriers_articles, 25/72"
#> [1] "Parsed drug_carriers_textbooks, 26/72"
#> [1] "Parsed drug_carriers_links, 27/72"
#> [1] "Parsed drug_carriers_polypeptides, 28/72"
#> [1] "Parsed drug_carriers_polypeptides_external_identifiers, 29/72"
#> [1] "Parsed drug_carriers_polypeptides_synonyms, 30/72"
#> [1] "Parsed drug_carriers_polypeptides_pfams, 31/72"
#> [1] "Parsed drug_carriers_polypeptides_go_classifiers, 32/72"
#> [1] "Parsed drug_carriers, 33/72"
#> [1] "Parsed drugs main attributes, 34/72"
#> [1] "Parsed drug_enzymes_actions, 35/72"
#> [1] "Parsed drug_enzymes_articles, 36/72"
#> [1] "Parsed drug_enzymes_textbooks, 37/72"
#> [1] "Parsed drug_enzymes_links, 38/72"
#> [1] "Parsed drug_enzymes_polypeptides, 39/72"
#> [1] "Parsed drug_enzymes_polypeptides_external_identifiers, 40/72"
#> [1] "Parsed drug_enzymes_polypeptides_synonyms, 41/72"
#> [1] "Parsed drug_enzymes_polypeptides_pfams, 42/72"
#> [1] "Parsed drug_enzymes_polypeptides_go_classifiers, 43/72"
#> [1] "Parsed drug_enzymes, 44/72"
#> [1] "Parsed drug_manufacturers, 45/72"
#> [1] "Parsed drug_pathway_enzyme, 46/72"
#> [1] "Parsed drug_pathway_drugs, 47/72"
#> [1] "Parsed drug_pathway, 48/72"
#> [1] "Parsed drug_prices, 49/72"
#> [1] "Parsed drug_reactions, 50/72"
#> [1] "Parsed drug_reactions_enzymes, 51/72"
#> [1] "Parsed drug_sequences, 52/72"
#> [1] "Parsed drug_targets_polypeptides_external_identifiers, 53/72"
#> [1] "Parsed drug_targets_polypeptides_synonyms, 54/72"
#> [1] "Parsed drug_targets_polypeptides_pfams, 55/72"
#> [1] "Parsed drug_targets_polypeptides_go_classifiers attributes, 56/72"
#> [1] "Parsed drug_targets_actions, 57/72"
#> [1] "Parsed drug_targets_articles, 58/72"
#> [1] "Parsed drug_targets_textbooks, 59/72"
#> [1] "Parsed drug_targets_links, 60/72"
#> [1] "Parsed drug_targets_polypeptide, 61/72"
#> [1] "Parsed drug_targets, 62/72"
#> [1] "Parsed drug_transporters_actions, 63/72"
#> [1] "Parsed drug_transporters_articles, 64/72"
#> [1] "Parsed drug_transporters_textbooks, 65/72"
#> [1] "Parsed drug_transporters_links, 66/72"
#> [1] "Parsed drug_transporters_polypeptides, 67/72"
#> [1] "Parsed drug_transporters_polypeptides_external_identifiers, 68/72"
#> [1] "Parsed drug_transporters_polypeptides_synonyms, 69/72"
#> [1] "Parsed drug_transporters_polypeptides_pfams, 70/72"
#> [1] "Parsed drugs_transporters_polypeptides_go_classifiers, 71/72"
#> [1] "Parsed drugs_transporters, 72/72"

#> $drugs
#> # A tibble: 0 x 0
#> 
#> $drug_groups
#> # A tibble: 0 x 0
#> 
#> $drug_articles
#> # A tibble: 0 x 0
#> 
#> $drug_books
#> # A tibble: 0 x 0
#> 
#> $drug_links
#> # A tibble: 0 x 0
#> 
#> $drug_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_products
#> # A tibble: 0 x 0
#> 
#> $drug_mixtures
#> # A tibble: 0 x 0
#> 
#> $drug_packagers
#> # A tibble: 0 x 0
#> 
#> $drug_categories
#> # A tibble: 0 x 0
#> 
#> $drug_affected_organisms
#> # A tibble: 0 x 0
#> 
#> $drug_dosages
#> # A tibble: 0 x 0
#> 
#> $drug_ahfs_codes
#> # A tibble: 0 x 0
#> 
#> $drug_pdb_entries
#> # A tibble: 0 x 0
#> 
#> $drug_patents
#> # A tibble: 0 x 0
#> 
#> $drug_food_interactions
#> # A tibble: 0 x 0
#> 
#> $drug_interactions
#> # A tibble: 0 x 0
#> 
#> $experimental_properties
#> # A tibble: 0 x 0
#> 
#> $drug_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_external_links
#> # A tibble: 0 x 0
#> 
#> $drug_snp_effects
#> # A tibble: 0 x 0
#> 
#> $drug_snp_adverse_drug_reactions
#> # A tibble: 0 x 0
#> 
#> $drug_atc_codes
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_actions
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_articles
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_links
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_carriers
#> # A tibble: 0 x 0
#> 
#> $classificationsparse_drug_classifications
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_actions
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_articles
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_links
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes
#> # A tibble: 0 x 0
#> 
#> $drug_manufacturers
#> # A tibble: 0 x 0
#> 
#> $drug_pathway_enzyme
#> # A tibble: 0 x 0
#> 
#> $drug_pathway_drugs
#> # A tibble: 0 x 0
#> 
#> $drug_pathway
#> # A tibble: 0 x 0
#> 
#> $drug_prices
#> # A tibble: 0 x 0
#> 
#> $drug_reactions
#> # A tibble: 0 x 0
#> 
#> $drug_reactions_enzymes
#> # A tibble: 0 x 0
#> 
#> $drug_sequences
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_targets_actions
#> # A tibble: 0 x 0
#> 
#> $drug_targets_articles
#> # A tibble: 0 x 0
#> 
#> $drug_targets_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_targets_links
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptide
#> # A tibble: 0 x 0
#> 
#> $drug_targets
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_actions
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_articles
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_links
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_transporters
#> # A tibble: 0 x 0
#> 

parse_drug_all(TRUE)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("NULL", "character", "tbl_df"), new("standardGeneric",     .Data = function (conn, name, value, ...)     standardGeneric("dbWriteTable"), generic = structure("dbWriteTable", package = "DBI"),     package = "DBI", group = list(), valueClass = character(0),     signature = c("conn", "name", "value"), default = NULL, skeleton = (function (conn,         name, value, ...)     stop("invalid call in method dispatch to 'dbWriteTable' (no default method)",         domain = NA))(conn, name, value, ...)), <environment>): unable to find an inherited method for function 'dbWriteTable' for signature '"NULL", "character", "tbl_df"'

parse_drug_all(save_table = FALSE)

#> [1] "Parsed drugs main attributes, 1/72"
#> [1] "Parsed drug_groups, 2/72"
#> [1] "Parsed drug_articles, 3/72"
#> [1] "Parsed drug_books, 4/72"
#> [1] "Parsed drug_links, 5/72"
#> [1] "Parsed drug_synonyms, 6/72"
#> [1] "Parsed drug_products, 7/72"
#> [1] "Parsed drug_mixtures, 8/72"
#> [1] "Parsed drug_packagers, 9/72"
#> [1] "Parsed drug_categories, 10/72"
#> [1] "Parsed drug_affected_organisms, 11/72"
#> [1] "Parsed drug_dosages, 12/72"
#> [1] "Parsed drug_ahfs_codes, 13/72"
#> [1] "Parsed drug_pdb_entries, 14/72"
#> [1] "Parsed drug_patents, 15/72"
#> [1] "Parsed drug_food_interactions, 16/72"
#> [1] "Parsed drug_interactions, 17/72"
#> [1] "Parsed drug_experimental_properties, 18/72"
#> [1] "Parsed drug_external_identifiers, 19/72"
#> [1] "Parsed drug_external_links, 20/72"
#> [1] "Parsed drug_snp_effects, 21/72"
#> [1] "Parsed drug_snp_adverse_drug_reactions, 22/72"
#> [1] "Parsed drug_atc_codes, 23/72"
#> [1] "Parsed drug_carriers_actions, 24/72"
#> [1] "Parsed drug_carriers_articles, 25/72"
#> [1] "Parsed drug_carriers_textbooks, 26/72"
#> [1] "Parsed drug_carriers_links, 27/72"
#> [1] "Parsed drug_carriers_polypeptides, 28/72"
#> [1] "Parsed drug_carriers_polypeptides_external_identifiers, 29/72"
#> [1] "Parsed drug_carriers_polypeptides_synonyms, 30/72"
#> [1] "Parsed drug_carriers_polypeptides_pfams, 31/72"
#> [1] "Parsed drug_carriers_polypeptides_go_classifiers, 32/72"
#> [1] "Parsed drug_carriers, 33/72"
#> [1] "Parsed drugs main attributes, 34/72"
#> [1] "Parsed drug_enzymes_actions, 35/72"
#> [1] "Parsed drug_enzymes_articles, 36/72"
#> [1] "Parsed drug_enzymes_textbooks, 37/72"
#> [1] "Parsed drug_enzymes_links, 38/72"
#> [1] "Parsed drug_enzymes_polypeptides, 39/72"
#> [1] "Parsed drug_enzymes_polypeptides_external_identifiers, 40/72"
#> [1] "Parsed drug_enzymes_polypeptides_synonyms, 41/72"
#> [1] "Parsed drug_enzymes_polypeptides_pfams, 42/72"
#> [1] "Parsed drug_enzymes_polypeptides_go_classifiers, 43/72"
#> [1] "Parsed drug_enzymes, 44/72"
#> [1] "Parsed drug_manufacturers, 45/72"
#> [1] "Parsed drug_pathway_enzyme, 46/72"
#> [1] "Parsed drug_pathway_drugs, 47/72"
#> [1] "Parsed drug_pathway, 48/72"
#> [1] "Parsed drug_prices, 49/72"
#> [1] "Parsed drug_reactions, 50/72"
#> [1] "Parsed drug_reactions_enzymes, 51/72"
#> [1] "Parsed drug_sequences, 52/72"
#> [1] "Parsed drug_targets_polypeptides_external_identifiers, 53/72"
#> [1] "Parsed drug_targets_polypeptides_synonyms, 54/72"
#> [1] "Parsed drug_targets_polypeptides_pfams, 55/72"
#> [1] "Parsed drug_targets_polypeptides_go_classifiers attributes, 56/72"
#> [1] "Parsed drug_targets_actions, 57/72"
#> [1] "Parsed drug_targets_articles, 58/72"
#> [1] "Parsed drug_targets_textbooks, 59/72"
#> [1] "Parsed drug_targets_links, 60/72"
#> [1] "Parsed drug_targets_polypeptide, 61/72"
#> [1] "Parsed drug_targets, 62/72"
#> [1] "Parsed drug_transporters_actions, 63/72"
#> [1] "Parsed drug_transporters_articles, 64/72"
#> [1] "Parsed drug_transporters_textbooks, 65/72"
#> [1] "Parsed drug_transporters_links, 66/72"
#> [1] "Parsed drug_transporters_polypeptides, 67/72"
#> [1] "Parsed drug_transporters_polypeptides_external_identifiers, 68/72"
#> [1] "Parsed drug_transporters_polypeptides_synonyms, 69/72"
#> [1] "Parsed drug_transporters_polypeptides_pfams, 70/72"
#> [1] "Parsed drugs_transporters_polypeptides_go_classifiers, 71/72"
#> [1] "Parsed drugs_transporters, 72/72"

#> $drugs
#> # A tibble: 0 x 0
#> 
#> $drug_groups
#> # A tibble: 0 x 0
#> 
#> $drug_articles
#> # A tibble: 0 x 0
#> 
#> $drug_books
#> # A tibble: 0 x 0
#> 
#> $drug_links
#> # A tibble: 0 x 0
#> 
#> $drug_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_products
#> # A tibble: 0 x 0
#> 
#> $drug_mixtures
#> # A tibble: 0 x 0
#> 
#> $drug_packagers
#> # A tibble: 0 x 0
#> 
#> $drug_categories
#> # A tibble: 0 x 0
#> 
#> $drug_affected_organisms
#> # A tibble: 0 x 0
#> 
#> $drug_dosages
#> # A tibble: 0 x 0
#> 
#> $drug_ahfs_codes
#> # A tibble: 0 x 0
#> 
#> $drug_pdb_entries
#> # A tibble: 0 x 0
#> 
#> $drug_patents
#> # A tibble: 0 x 0
#> 
#> $drug_food_interactions
#> # A tibble: 0 x 0
#> 
#> $drug_interactions
#> # A tibble: 0 x 0
#> 
#> $experimental_properties
#> # A tibble: 0 x 0
#> 
#> $drug_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_external_links
#> # A tibble: 0 x 0
#> 
#> $drug_snp_effects
#> # A tibble: 0 x 0
#> 
#> $drug_snp_adverse_drug_reactions
#> # A tibble: 0 x 0
#> 
#> $drug_atc_codes
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_actions
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_articles
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_links
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_carriers_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_carriers
#> # A tibble: 0 x 0
#> 
#> $classificationsparse_drug_classifications
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_actions
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_articles
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_links
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_enzymes
#> # A tibble: 0 x 0
#> 
#> $drug_manufacturers
#> # A tibble: 0 x 0
#> 
#> $drug_pathway_enzyme
#> # A tibble: 0 x 0
#> 
#> $drug_pathway_drugs
#> # A tibble: 0 x 0
#> 
#> $drug_pathway
#> # A tibble: 0 x 0
#> 
#> $drug_prices
#> # A tibble: 0 x 0
#> 
#> $drug_reactions
#> # A tibble: 0 x 0
#> 
#> $drug_reactions_enzymes
#> # A tibble: 0 x 0
#> 
#> $drug_sequences
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_targets_actions
#> # A tibble: 0 x 0
#> 
#> $drug_targets_articles
#> # A tibble: 0 x 0
#> 
#> $drug_targets_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_targets_links
#> # A tibble: 0 x 0
#> 
#> $drug_targets_polypeptide
#> # A tibble: 0 x 0
#> 
#> $drug_targets
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_actions
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_articles
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_textbooks
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_links
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_external_identifiers
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_synonyms
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_pfams
#> # A tibble: 0 x 0
#> 
#> $drug_transporters_polypeptides_go_classifiers
#> # A tibble: 0 x 0
#> 
#> $drug_transporters
#> # A tibble: 0 x 0
#>